Work around a side-effect of the GEOM gpart command.
If we are in the GEOM world (FBSD8+) and we do not find the 4th partition
it may be because a previous gpart call noticed the unused partition
and deleted it. This happens in elabinelab where we first repurpose
the second (Linux) partition.

If this happens, we just try re-creating the 4th partition. gpart
will "do the right thing" in terms of offset and size.