Change the DHCP configuration code to allow for IPs on the actual control network, so that we can use routable IPs.
Create a swap volume for the guest. This addition is reflected in
rc.frisbee but not the linux version; looks for a disk with a dos
partition labeled as linux swap.

Look in the node attributes for additonal disk specifications and
create those disks (logical volumes). This is currently used in the
XEN elabinelab to add more disks for inner boss and ops, and look
like:

  $myboss add-attribute XEN_EXTRADISKS "disk1:10G,disk2:10G,disk3:10G"
  $myops  add-attribute XEN_EXTRADISKS "disk1:4G,disk2:10G"

It is up to the guest to do something with these, we do not partition
or label the volumes.