Add a VGAONLY attribute to the post-frisbee customization.
This is for BSD, so that we can tweak the default /boot/loader.conf file
to turn off "comconsole" on machines that don't have a UART.
In my experience, trying to probe the UART on newer machines that don't
have one causes a kernel crash.

Currently this attribute is just signalled by the existance of the
/etc/testbed/isvgaonly file in the diskloader MFS. But this will be
eventually sent over via "tmcc loadinfo" so that we can set the attribute
on a per-node-type or per-node basis.