Couple of utility scripts to capture and restore XEN vms so that we can move them elsewhere.
	capturevm.pl pcvmxxx-yy boss

will create a directory called boss and take imagezips of all the lvms
into that directory. A slightly modified xen config is written to the
directory. The XEN VM should not be running of course, and the easiest
way to stop it (without losing the disks) is:

	/usr/local/etc/emulab/vnodesetup -jh pcvmxxx-yy

Copy the directory someplace, and then run:

	restorevm.pl boss /path/to/above/stuff

which create new LVMs and then imageunzip the contents into them. A
slightly modified xen config file is then written to /var/tmp, which
can started with "xm create /var/tmp/boss.conf".

No rocket science here, but this was the easiest way to get all the
disks we create for XEN elabinelab, and since they can pretty big, we
want to use imagezip.