Improve cross referencing between geni-cm and emulab datbases.
Add a datetime form to the shownodehistory we page so that a testbed
admin can plug in a specific date, and find out what that node was
doing at the time. Changes in the backend (node_history script) to
support this. Note that the table is hard to seach for such a case,
and so need to let node_history do its thing and then port process the
records list. Unfortunately, the timestamps are unsigned ints, but
perl does not handle those properly, so had to pull in Math::BigInt to
deal with it.

On the output page, include a link to the genihistory page if a node
was part of a slice.

On the genihistory page, add a new argument, slice_uuid, to look for
the records for a specific slice.