Add a NODECONSOLE defs* variable.
This is not as general or as useful as you might think.
Right now it simply controls what device gets used as console when
customizing the MFSes. "sio" is the default. Other choices correspond
to pxeboot variants: sio2, vga, null. Choosing vga also sets the magic
VGAONLY setting in the frisbee MFS ensuring that any customized (via slicefix)
FreeBSD image doesn't try to use the serial port as console.

A real solution for choosing node consoles would not be per-testbed.
It would be per-node-type and per-node. Right now the console type is
selected in that fashion via a combination of the pxe_boot_path and
frisbee/admin MFS OSIDs. At some point we should make the console setting
explicit.