Add a node-default pxe_boot_path to go along with node-type-default.
As if things weren't hairy enough for the pxe_boot_path, we need a per-node
default value. This should not be confused with the next_pxe_boot_path or
pxe_boot_path fields in the nodes table, which is a per-node value but only
for the lifetime of the current experiment.

We need this new field specifically so that we can switch some, but not all,
of the d710s over to the TPM-aware pxeboot.

For the record, the select order for setting the filename field in
dhcpd.conf is:

 1. nodes.next_pxe_boot_path for node
 2. nodes.pxe_boot_path for node
 3. "pxe_boot_path" node_attribute for node
 4. "pxe_boot_path" node_type_attribute for node type
 5. don't set (i.e., use the dhcpd.conf global default).