Print a less obtuse error if they attempt to image an "empty" partition.
This is in the node type check where we look at what is currently in the
partition they are imaging, and base the set of valid node type on that.
However, if they specify a partition that doesn't currently have an image,
we would previously effectively tell them that the image will not run on
the node type that it is currently running on!

Now we just say "there is no current image in that partition."