Pass through bootinfo flags on tmcc "bootwhat" command.
bootwhat will now return a FLAGS=%d value corresponding to the flags
field in the boot_what struct.

NOTE: THIS REQUIRED A TMCD VERSION BUMP. We are now at version 35.
The issue was backward compatibility with existing CD/dongle boot images
which are overly strict in their parsing of the returned bootwhat values.

Added a new boot_what flag (the whole point of this) to signify if the
entity being returned is part of the "secure boot" path. This is used
by the gPXE dongle to determine whether it needs to do a trusted boot
path "sign-off" for the MFS it downloads. We used to use the name of
the MFS as our heuristic for this.

bootinfo uses the new tbdb.os_info osfeature "ontrustedboot" to determine
whether to set the flag.