Add LINUX_FSNODE support contributed by Maarten Dewispelaere at intec.ugent.be
These are not complete changes, but enough to support the usage these
guys need. In this case, I think it is a hand crafted linux node that
allows boss to ssh in and do the exports setup stuff. But hey, it is a
start on actually supporting a linux FS node.