Comment out TPM-aware "imagekey" call.
This call is already done in script1.sh. My intention in putting it in
rc.frisbee was to phase out script[12].sh. But there is enough magic in
those scripts that I am just going to leave them be. In particular is the
constant restarting of tcsd to avoid hangs.