Hack to shutup perl 5.12 about unused SOUT variable.
The perl warnings can screw up some of our scripts which expect certain
output; e.g., users of "setsitevar -c" expect to see the value of a sitevar
and not some nagging perl message.

Don't want to remove SOUT altogether since I don't know how that will
affect the regular libtblog (which includes libtblog_simple). So I just
threw in an "if (0)" which prints to SOUT...