Fix another cause of trafgen not building.
This might have been fixed in later gmakes, but at least for the circa-FBSD6
version, the shell PWD variable is not correctly set when a shell script is
run from a make target via:
    $(MAKE) -C foo target
instead of:
    cd foo && $(MAKE) target
In the former case, $PWD would not reflect that it was in "foo".
The upshot is that the tg2.0 tarball was being extracted in the "event"
directory rather than "event/trafgen" so the make rules could not find it.
The same basic shell script worked for iperf and rude because they were
invoked via the latter form.

Anyway, now we just use dir=`pwd` instead of dir=$PWD.