Add -L option to capture to print out how long it has been since last output.
We currently put out a time stamp when we detect output after being idle for
a long time. But the time stamp is the time at which we started getting output
again, so it doesn't give us any idea how long it was idle (i.e., when the
last previous output happened).

So -L changes the format of the stamp from "STAMP{ctime-string}" to
"STAMP{idle@ctime-string}", where "idle" is the number of seconds it has been
since the last output.