Remove dependency on a racy state transition.
The recent reboot-node-if-going-from-PXEWAIT-to-SECURELOAD change to
bootinfo was checking that the node was in PXEKERNEL/PXEBOOTING. However,
the transition from PXEWAKEUP->PXEBOOTING may not have happened yet.

However, the check for PXEBOOTING was unnecessary anyway, since bootinfo
is the one that sends that state and it *always* does it just before calling
query_bootinfo_db (the guy relying on the racy state change). Thus, the node
is intended to be in PXEBOOTING even if that hasn't yet been recorded in
the DB. So just remove the state check (we still check the op_mode).

Note that this only affects nodes that are using the secure diskload MFS.