Add host and path headers to gitmail
For now, these go into headers (X-Git-Repo-Host and X-Git-Repo-Path)
Might consider moving them to the body of the mail at some point.

Note that I *don't* attempt to figure out the actual 'URL' for the
repo, since that is *not* necessarily automatically derivable: eg.
it might be any of ssh,git, or http protocols, and the path (esp.
for git: and http:) might not match the path in the filesystem. So
this suggests that wat goes into the body of the mail ought to be
set with a config variable.