Fixes to libvnode_openvz for Fedora 15
Set the hostname for the vnode using vzctl to make sure it gets set
correctly.  Otherwise the vnode will inherit the physical host's
hostname, and since this is not "localhost" or "localhost.localdomain"
the networking scripts inside the container will not set reset it to
the value in /etc/sysconfig/network.

Use the "inside" MAC address rather than the "outside" one when
creating macvlan devices.

Only try to delete interfaces if they still exist when tearing down
the vnode.  The kernel itself automatically deletes "virtual"
interfaces, which includes imq and macvlan.