Define some additional configure vars in the clientside configure.
Against my better judgement, use configure to set some vars in clientside
scripts. I try to avoid this to keep our client images from becoming Utah
specific.

This is specifically for run_linktest.pl.in, which is installed on both
ops and nodes. I could fix it to not use configure vars on nodes, but that
would break its primary use on ops.

OTOH, I fixed elab_linktest.pl.in to not use configure variables, since
it only runs on nodes.