Add ReadHistoryRecords() API to download the aggregate history from the CH database.
Requires a credential (possibly delegated) with the "readhistory"
privilege.

   ReadHistoryRecords(credential, int index, int count);

where index and count are optional, but to be useful you will want to
track the index and ask for "count" records at a time until you get
back an error saying there are no more to give you. You are limited to
max 100 (the default is 10) time at a time.

The return value is a list of records, where each record is an array
of the data in the aggregate_history table and the manifest. There
might be multiple records if the slice was updated, since there will
be multiple manifests. I won't describe it, you can just print out a
record and see it, it is all pretty obvious.