Initial support for GENI portal, switched to OpenStreetMaps, and bug fixes
Flack can now be bootstrapped according to the requirements of the GENI portal.  'flackportal.html' is provided as an example.  First, the server certs and client cert/key pair are set in JavaScript before launching Flack.  Next, several variables are handed to Flack through URL parameters.

Very important and of note is that Flack has finally switched from using Google Maps (deprecated for some time and requires domain-specific keys) to OpenStreetMaps which doesn't require a key.  This is important for other websites who want to host Flack (like the GENI portal).