Fix libosload to properly get the size of images it cannot read directly.
If a project image has been "exported" with grantimage, os_load may not
be able to stat it due to unix permissions. So have os_load make a query
to the local frisbee master server to get attributes for the image. The
master server already knows how to deal with these exported images.
This query also works for an inner boss determining the size of an image
it has not yet downloaded and thus subsumes that case.