Fix a capserver vulnerability reported by John Hickey.
Validate those SQL args!

NOTE: we also ensure that the reporting node is listed as a legit tip server
in the tipservers table. This means that capture may stop working on nodes
whose servers are not in the table! SQL update 291 will add any servers
listed in tiplines entries that are not in tipservers to prevent this breakage.