Return VGAONLY in the loadinfo clause, if set in the node_type_attributes or the node_attributes table.
Note that I had to fix the query; the original select required that
the key be in the node_type_attributes table, else it would fail to
find it in the node_attributes table. In other words, I could not put
vgaonly on a specific node unless there was a definition in the
node_type_attributes table.

I've changed the query to use a "union" of two selects; this query is
intended to select certain attributes from node_type_attributes table,
but allow them to be overridden by entries in the node_attributes
table for the node making the request. The "union" of two selects,
allow results from the second select on node_attributes to overwrite
anything returned for the same key in the first select on
node_type_attributes.