Add '-R' option to assign
When set, randomizes the order of nodes before making pclasses. This
has the effect of putting different nodes at the 'front' of the pclass,
The idea is to avoid 'hot spots' of nodes that get over-used simply
because they happen to appear earlier in the .ptop file.

NB: because STL seems to use a different source of randomness, using
this with '-s' to set a seed does *not* make assign deterministic.