Add reverse DNS lookup for the jail network.
The GPO wants this for the protogeni racks. We now build reverse
map files for the 172.16 subnet, although we do it on a /16 boundry
to avoid a zillion zone files.

I am not planning to write an update script for this, since it would
require scripting changes to named.conf, which I am loath to do. So I
will do it by hand in Utah, and new sites (racks) will get it. If a
site wants it:

	boss> cd obj/named

Copy all of the 172 files to /etc/named/reverse
Copy all of the 172 zone entries from named.conf to /etc/named/named.conf

	boss> named_setup