Bug fix for fixing VMs to nodes. Changes to shared vlans.
People try to fix pc433 to pc433. The former is a node in the topo,
the later is a physical node. Causes confusion, breaks. Look for
this corner case.

Allow lans that use shared lans, to have more then one port. So now
you can do this in your rspec:

   <link client_id="link0">
     <vlan:link_shared_vlan name="openflow-mesoscale" />
     <interface_ref client_id="node1:if" />
     <interface_ref client_id="node2:if" />
   </link>