Modifications to linktest for Windows.
Two things changed:

* Always prime Windows' ARP cache with a quick ping.

The first ping under Windows is often wildly off compared to the rest, which
throws off the average enough to fail the latency check.

* Lower the upper bandwidth limit for bandwidth test

iperf under Win7 can't push over 40Mbps, at least not reliably, with the
parameters we use.  Prevent BW testing from being performed on Windows if
it is over this.  To be documented as a caveat, and may revisit later.