Minor tweak to "waitmode" in libreboot.
Previously ISUP and TBFAILED were the two states that signified that a
node had rebooted to the satisfaction of libreboot (with waitmode==1).
Add RELOAD/RELOADING to that list since the frisbee MFS never sends either
of ISUP or TBFAILED.

Required a modification to TBNodeStateWait() to allow waiting for an
op-mode/state combo as well as just a state.

I made this change in anticipation that it would be useful for more
responsive monitoring of failure in tbswap/os_load. But now I am no so sure.