Fixes for Win7: route setup, os-specific rc
Rework the OS specific rc script a bit - still awful to behold, but a bit
better.  Add a new bit of magic to capture the mapping from mac address
to internal windows interface index.  May use this later for pinning an
route to the correct interface in rc.ifc.

Create a couple of symlinks for to getmac and devcon utilities from
the Makefile.

Add a delay after configuring the IP on an interface (before adding routes).
Windows can apparently return from the netsh configuration command before
everything is done.