Call init() method in makeSortable() to fix broken table sorting.
This is needed due to the odd way in which we use sorttable.
In normal use, sorttable.init would be called via the DOMContentLoaded
event hook and this would invoke makeSortable() to do the DOM magic
necessary to ensure all tables marked sortable are make sortable.

But due to our use of AJAX-y dynamic content loading, this won't work
(or so I am told!) So we generate our own call to makeSortable(mytable)
when we produce any sortable table "mytable". But if init() has not been
called before that first call to makeSortable, the DATE_RE regex will
not be initialized and strange things happen in guessType and consequently,
in the sort functions themselves.