Make sorttable_alpha case-insensitive and use stable sort.
The former ensures that we get the same behavior as when the same
sorted output is produced using "order by" in mysql. I added a
sortable_alpha_cs that we can use for a column should we discover
a case where we really want a case-sensitive sort.

The latter ensures that a user can at least manually do a multi-column
sort (e.g., the equiv of "order by state,eid" when displaying experiments
by clicking on the eid column and then on the state column).