Deal with failure to start up listeners more gracefully.
Fix a weird problem where the failure to start up the "crude" listener
got reflected as a failure of the latency test. This was happening on
Fedora 15 because we could not run crude at elevated priority.

Part of the fix is to wait a second after starting a listener and then
make sure it is running. In the case of such an early failure of "crude",
we then try again without elevated priority.